refactor!: Move slack application under api.
