+++ web/api/index_handler.py
@@ -14,6 +14,8 @@ class name: route replace capitalise route segments remove `/` and `_`. Example:
 """
 
 # for now we'll manually add imports. TODO: convert to walk the directory and dynamically import using importlib
+import importlib
+
 from . import (
     chat_completion_handler,  # noqa: F401 DO NOT REMOVE
     hello_handler,  # noqa: F401 DO NOT REMOVE
@@ -22,3 +24,5 @@ from . import (
     threads_handler,  # noqa: F401 DO NOT REMOVE
     token_handler,  # noqa: F401 DO NOT REMOVE
 )
+
+importlib.import_module("web.api.integration.slack.index_handler")
